refactor : api 리펙토링
