+++ backend/src/main/java/com/runinto/config/webSocketConfig.java
@@ -1,6 +1,6 @@
 package com.runinto.config;
 
-import com.runinto.chat.domain.CustomWebSocketHandler;
+import com.runinto.chat.presntation.CustomWebSocketHandler;
 import com.runinto.util.UserHandshakeInterceptor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.socket.config.annotation.EnableWebSocket;
